refactor: extract vminstanceutils.go from vminstance.go and cloudprovider.go
Move VM utility functions into a dedicated vminstanceutils.go file:
- Priority maps (KarpCapacityTypeToVMPriority, VMPriorityToKarpCapacityType, VMPriorityToScaleSetPriority)
- Extension-related constants and helpers (aksIdentifyingExtensionName, GetManagedExtensionNames, isAKSIdentifyingExtensionEnabled)
- ErrorCodeForMetrics, ConvertToVirtualMachineIdentity, GetCapacityTypeFromVM, GetScaleSetPriorityLabelFromVM

Move vmInstanceToNodeClaim from CloudProvider to BuildNodeClaimFromVM in instance package,
paralleling BuildNodeClaimFromAKSMachine. Move GetNodeClaimNameFromVMName to instance package.
