+++ pkg/cloudprovider/suite_drift_test.go
@@ -36,6 +36,7 @@ import (
 	"github.com/Azure/karpenter-provider-azure/pkg/apis/v1beta1"
 	"github.com/Azure/karpenter-provider-azure/pkg/fake"
 	"github.com/Azure/karpenter-provider-azure/pkg/operator/options"
+	"github.com/Azure/karpenter-provider-azure/pkg/providers/instance"
 	"github.com/Azure/karpenter-provider-azure/pkg/test"
 	. "github.com/Azure/karpenter-provider-azure/pkg/test/expectations"
 )
@@ -352,7 +353,7 @@ var _ = Describe("CloudProvider", func() {
 				Expect(azureEnv.VirtualMachinesAPI.VirtualMachineCreateOrUpdateBehavior.CalledWithInput.Len()).To(Equal(1))
 				input := azureEnv.VirtualMachinesAPI.VirtualMachineCreateOrUpdateBehavior.CalledWithInput.Pop()
 				// Corresponding NodeClaim - fetch from API server so it has all status fields (including ImageID)
-				nodeClaimName := GetNodeClaimNameFromVMName(input.VMName)
+				nodeClaimName := instance.GetNodeClaimNameFromVMName(input.VMName)
 				driftNodeClaim = &karpv1.NodeClaim{}
 				Expect(env.Client.Get(ctx, types.NamespacedName{Name: nodeClaimName}, driftNodeClaim)).To(Succeed())
 				// ExpectProvisioned doesn't set Status.NodeName
