+++ pkg/providers/instance/vminstance.go
@@ -31,7 +31,6 @@ import (
 	"github.com/samber/lo"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
-	azureclouds "sigs.k8s.io/cloud-provider-azure/pkg/azclient"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 	karpv1 "sigs.k8s.io/karpenter/pkg/apis/v1"
 	corecloudprovider "sigs.k8s.io/karpenter/pkg/cloudprovider"
@@ -54,71 +53,6 @@ import (
 	"github.com/Azure/karpenter-provider-azure/pkg/providers/networksecuritygroup"
 	"github.com/Azure/karpenter-provider-azure/pkg/utils"
 )
-
-var (
-	KarpCapacityTypeToVMPriority = map[string]armcompute.VirtualMachinePriorityTypes{
-		karpv1.CapacityTypeSpot:     armcompute.VirtualMachinePriorityTypesSpot,
-		karpv1.CapacityTypeOnDemand: armcompute.VirtualMachinePriorityTypesRegular,
-	}
-	VMPriorityToKarpCapacityType = map[armcompute.VirtualMachinePriorityTypes]string{
-		armcompute.VirtualMachinePriorityTypesSpot:    karpv1.CapacityTypeSpot,
-		armcompute.VirtualMachinePriorityTypesRegular: karpv1.CapacityTypeOnDemand,
-	}
-	// Note that there is no ScaleSetPriorityToKarpCapacityType because the karpenter.sh/capacity-type
-	// label is the "official" label that we actually key priority off of. Selection still works though
-	// because when we list instance types on-demand offerings always have v1beta1.ScaleSetPriorityRegular
-	// and spot instances always have v1beta1.ScaleSetPrioritySpot, so the correct karpenter.sh/capacity-type
-	// label is still selected even if the user is using kubernetes.azure.com/scalesetpriority only on the NodePool.
-	VMPriorityToScaleSetPriority = map[armcompute.VirtualMachinePriorityTypes]string{
-		armcompute.VirtualMachinePriorityTypesSpot:    v1beta1.ScaleSetPrioritySpot,
-		armcompute.VirtualMachinePriorityTypesRegular: v1beta1.ScaleSetPriorityRegular,
-	}
-
-	aksIdentifyingExtensionEnvs = sets.New(
-		azureclouds.PublicCloud.Name,
-		azureclouds.ChinaCloud.Name,
-		azureclouds.USGovernmentCloud.Name,
-	)
-)
-
-const (
-	aksIdentifyingExtensionName = "computeAksLinuxBilling"
-	// TODO: Why bother with a different CSE name for Windows?
-	cseNameWindows = "windows-cse-agent-karpenter"
-	cseNameLinux   = "cse-agent-karpenter"
-)
-
-// ErrorCodeForMetrics extracts a stable Azure error code for metric labeling when possible.
-func ErrorCodeForMetrics(err error) string {
-	if err == nil {
-		return "UnknownError"
-	}
-	if azErr := sdkerrors.IsResponseError(err); azErr != nil {
-		if azErr.ErrorCode != "" {
-			return azErr.ErrorCode
-		}
-		return "UnknownError"
-	}
-	return "UnknownError"
-}
-
-// GetManagedExtensionNames gets the names of the VM extensions managed by Karpenter.
-// This is a set of 1 or 2 extensions (depending on provisionMode): aksIdentifyingExtension and (sometimes) cse.
-func GetManagedExtensionNames(provisionMode string, env *auth.Environment) []string {
-	var result []string
-	// Only including AKS identifying extension in the clouds it is supported in
-	if isAKSIdentifyingExtensionEnabled(env) {
-		result = append(result, aksIdentifyingExtensionName)
-	}
-	if provisionMode == consts.ProvisionModeBootstrappingClient {
-		result = append(result, cseNameLinux) // TODO: Windows
-	}
-	return result
-}
-
-func isAKSIdentifyingExtensionEnabled(env *auth.Environment) bool {
-	return aksIdentifyingExtensionEnvs.Has(env.Environment.Name)
-}
 
 type Resource = map[string]interface{}
 
@@ -1060,37 +994,4 @@ func (p *DefaultVMProvider) getCSExtension(cse string, isWindows bool, tags map[
 		},
 		Tags: tags,
 	}
-}
-
-func ConvertToVirtualMachineIdentity(nodeIdentities []string) *armcompute.VirtualMachineIdentity {
-	var identity *armcompute.VirtualMachineIdentity
-	if len(nodeIdentities) > 0 {
-		identityMap := make(map[string]*armcompute.UserAssignedIdentitiesValue)
-		for _, identityID := range nodeIdentities {
-			identityMap[identityID] = &armcompute.UserAssignedIdentitiesValue{}
-		}
-
-		if len(identityMap) > 0 {
-			identity = &armcompute.VirtualMachineIdentity{
-				Type:                   lo.ToPtr(armcompute.ResourceIdentityTypeUserAssigned),
-				UserAssignedIdentities: identityMap,
-			}
-		}
-	}
-
-	return identity
-}
-
-func GetCapacityTypeFromVM(vm *armcompute.VirtualMachine) string {
-	if vm != nil && vm.Properties != nil && vm.Properties.Priority != nil {
-		return VMPriorityToKarpCapacityType[*vm.Properties.Priority]
-	}
-	return ""
-}
-
-func GetScaleSetPriorityLabelFromVM(vm *armcompute.VirtualMachine) string {
-	if vm != nil && vm.Properties != nil && vm.Properties.Priority != nil {
-		return VMPriorityToScaleSetPriority[*vm.Properties.Priority]
-	}
-	return ""
 }
